pyproject.toml
[tool.poetry]
name = "twitter-clone-py"
version = "0.1.0"
description = ""
authors = ["toweringcloud <toweringcloud@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
